rename dictionary decoder
